[refactor] 로직 리팩토링 및 파일 이동
